encode return value as integer instead of string
The in-toto python reference implementation encodes the return-value
field of our link files as integer. Our specification specified this
field as string value. This commit fixes this inconsistency and
assigns the right data type to the return-value field.

Furthermore we remove a "return-value" line, outside of
of an actual byproducts object. May has been a remainder
of the signed part.